+++ src/TreeHouse/QueueBundle/Consumer/Limiter/LimitReachedException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace TreeHouse\QueueBundle\Consumer\Limiter;
+
+class LimitReachedException extends \RuntimeException
+{
+    /**
+     * @param string $message
+     *
+     * @return LimitReachedException
+     */
+    public static function withMessage($message)
+    {
+        return new self($message);
+    }
+}