+++ src/TreeHouse/QueueBundle/Consumer/Limiter/LimiterInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace TreeHouse\QueueBundle\Consumer\Limiter;
+
+use TreeHouse\QueueBundle\Consumer\Consumer;
+
+interface LimiterInterface
+{
+    /**
+     * @param Consumer $consumer
+     *
+     * @throws LimitReachedException
+     */
+    public function limitReached(Consumer $consumer);
+}